+++ CHANGELOG.md
@@ -4,6 +4,7 @@ ChangeLog
 
 ##### Changes for Android
 - Added [onlyAlertOnce](https://developer.android.com/reference/android/app/Notification.Builder#setOnlyAlertOnce(boolean)) option
+- Bugfix: Use app name as tag when calling NotificationManager.cancel
 
 #### Version 1.0.0 (17.08.2024)
 This Release contains mainly changes and fixes for the Android platform.